# internal/api/client/api/swagger.yaml
openapi: 3.0.3
info:
  title: Frank Auth Server API
  description: "Frank is a comprehensive authentication and authorization server that\
    \ provides OAuth2, passwordless login, MFA, \npasskeys, SSO, enterprise SSO, webhooks,\
    \ organizations, and API keys for machine-to-machine authentication.\n"
  contact:
    name: Frank Auth Server
    url: https://github.com/juicycleff/frank
  version: 1.0.0
servers:
- url: https://auth.example.com/api/v1
  description: Production server
- url: http://localhost:8080/api/v1
  description: Local development server
...
paths:
...
  /webhooks:
    get:
      tags:
      - Webhooks
      summary: List webhooks
      description: List webhooks with pagination
      operationId: listWebhooks
      parameters:
      - name: offset
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: integer
          default: 0
      - name: limit
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: integer
          default: 20
      - name: event_types
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: List of webhooks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_29"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
    post:
      tags:
      - Webhooks
      summary: Create webhook
      description: Create a new webhook
      operationId: createWebhook
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/webhooks_body"
        required: true
      responses:
        "201":
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
  /webhooks/{id}:
    get:
      tags:
      - Webhooks
      summary: Get webhook
      description: Get a specific webhook
      operationId: getWebhook
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Webhook details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
    put:
      tags:
      - Webhooks
      summary: Update webhook
      description: Update a specific webhook
      operationId: updateWebhook
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/webhooks_id_body"
        required: true
      responses:
        "200":
          description: Webhook updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Webhooks
      summary: Delete webhook
      description: Delete a specific webhook
      operationId: deleteWebhook
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "204":
          description: Webhook deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
  /webhooks/{id}/events:
    get:
      tags:
      - Webhooks
      summary: List webhook events
      description: List events for a specific webhook
      operationId: listWebhookEvents
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: offset
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: integer
          default: 0
      - name: limit
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: integer
          default: 20
      - name: event_type
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: delivered
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: boolean
      responses:
        "200":
          description: List of webhook events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_30"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
  /webhooks/{id}/events/{eventId}/replay:
    post:
      tags:
      - Webhooks
      summary: Replay webhook event
      description: Replay a specific webhook event
      operationId: replayWebhookEvent
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: eventId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Event replayed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEvent"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Webhook or event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
  /webhooks/trigger:
    post:
      tags:
      - Webhooks
      summary: Trigger webhook event
      description: Manually trigger a webhook event
      operationId: triggerWebhookEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/webhooks_trigger_body"
        required: true
      responses:
        "200":
          description: Event triggered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEvent"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/Error_error"
...
    Webhook:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        url:
          type: string
          format: uri
        event_types:
          type: array
          items:
            type: string
        active:
          type: boolean
        secret:
          type: string
        retry_count:
          type: integer
        timeout_ms:
          type: integer
        format:
          type: string
          enum:
          - json
          - form
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        organization_id:
          type: string
        metadata:
          type: object
          additionalProperties: true
      example:
        metadata: ""
        format: json
        active: true
        created_at: 2000-01-23T04:56:07.000+00:00
        secret: secret
        url: http://example.com/aeiou
        event_types:
        - event_types
        - event_types
        updated_at: 2000-01-23T04:56:07.000+00:00
        retry_count: 0
        organization_id: organization_id
        name: name
        id: id
        timeout_ms: 6
    WebhookEvent:
      type: object
      properties:
        id:
          type: string
        webhook_id:
          type: string
        event_type:
          type: string
        payload:
          type: object
          additionalProperties: true
        delivered:
          type: boolean
        status_code:
          type: integer
        attempts:
          type: integer
        last_attempt_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        response:
          type: string
        error:
          type: string
      example:
        event_type: event_type
        status_code: 0
        webhook_id: webhook_id
        payload: ""
        response: response
        last_attempt_at: 2000-01-23T04:56:07.000+00:00
        created_at: 2000-01-23T04:56:07.000+00:00
        delivered: true
        id: id
        error: error
        attempts: 6
...
    inline_response_200_29:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Webhook"
        total:
          type: integer
        pagination:
          $ref: "#/components/schemas/inline_response_200_22_pagination"
      example:
        total: 1
        pagination:
          total: 5
          offset: 6
          limit: 1
        data:
        - metadata: ""
          format: json
          active: true
          created_at: 2000-01-23T04:56:07.000+00:00
          secret: secret
          url: http://example.com/aeiou
          event_types:
          - event_types
          - event_types
          updated_at: 2000-01-23T04:56:07.000+00:00
          retry_count: 0
          organization_id: organization_id
          name: name
          id: id
          timeout_ms: 6
        - metadata: ""
          format: json
          active: true
          created_at: 2000-01-23T04:56:07.000+00:00
          secret: secret
          url: http://example.com/aeiou
          event_types:
          - event_types
          - event_types
          updated_at: 2000-01-23T04:56:07.000+00:00
          retry_count: 0
          organization_id: organization_id
          name: name
          id: id
          timeout_ms: 6
    webhooks_body:
      required:
      - event_types
      - name
      - url
      type: object
      properties:
        name:
          type: string
        url:
          type: string
          format: uri
        event_types:
          type: array
          items:
            type: string
        retry_count:
          type: integer
        timeout_ms:
          type: integer
        format:
          type: string
          default: json
          enum:
          - json
          - form
        metadata:
          type: object
          additionalProperties: true
    webhooks_id_body:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
          format: uri
        active:
          type: boolean
        event_types:
          type: array
          items:
            type: string
        retry_count:
          type: integer
        timeout_ms:
          type: integer
        format:
          type: string
          enum:
          - json
          - form
        metadata:
          type: object
          additionalProperties: true
    inline_response_200_30:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/WebhookEvent"
        total:
          type: integer
        pagination:
          $ref: "#/components/schemas/inline_response_200_22_pagination"
      example:
        total: 1
        pagination:
          total: 5
          offset: 6
          limit: 1
        data:
        - event_type: event_type
          status_code: 0
          webhook_id: webhook_id
          payload: ""
          response: response
          last_attempt_at: 2000-01-23T04:56:07.000+00:00
          created_at: 2000-01-23T04:56:07.000+00:00
          delivered: true
          id: id
          error: error
          attempts: 6
        - event_type: event_type
          status_code: 0
          webhook_id: webhook_id
          payload: ""
          response: response
          last_attempt_at: 2000-01-23T04:56:07.000+00:00
          created_at: 2000-01-23T04:56:07.000+00:00
          delivered: true
          id: id
          error: error
          attempts: 6
    webhooks_trigger_body:
      required:
      - event_type
      - payload
      type: object
      properties:
        event_type:
          type: string
        payload:
          type: object
          additionalProperties: true
        headers:
          type: object
          additionalProperties:
            type: string
...
    Error_error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: object
          additionalProperties: true
    inline_response_200_22_pagination:
      type: object
      properties:
        offset:
          type: integer
        limit:
          type: integer
        total:
          type: integer
      example:
        total: 5
        offset: 6
        limit: 1
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT token obtained from the /auth/login endpoint
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      description: API key for machine-to-machine authentication
      name: X-API-Key
      in: header